renaming wakuv2 for waku
